Fix for Matchplay Ratings 1250 bug; added players
Some ratings changed from 1 to 2, or 1 to 3, because of the bug that gave default Matchplay Ratings at 1250. Also looked up and added new subs from weeks 1 and 2 of season 12. And fixed Chole Crumbliss spelling, and Laura Minter who is still Laura Bodine in IFPA (she needs to change that with them).

Did not add "Kitty" who subbed in week one for JMF because I don't know who or what that is. ;-) If they reappear we need to get an actual name for them.